code/nnv/examples/Submission/FORMALISE2025/requirements.txt
coloredlogs==15.0.1
contourpy==1.2.1
cycler==0.12.1
eagerpy==0.30.0
flatbuffers==24.3.25
fonttools==4.53.0
humanfriendly==10.0
kiwisolver==1.4.5
matlabengine==24.1.2
matplotlib==3.9.0
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.18.0
packaging==24.1
pillow==10.3.0
protobuf==5.27.1
pyparsing==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
scipy==1.14.1
six==1.16.0
sympy==1.12.1
typing_extensions==4.12.2
urllib3==2.2.3
